+++ CHANGELOG.rst
@@ -1,13 +1,14 @@
 Changelog
 =========
 
-0.4.0 (2022-01-11)
+0.4.0 (2022-01-12)
 ------------------
 
 - Recognize builtin exceptions as not having slots.
 - Split ``--exclude-modules`` and ``exclude-classes``.
 - Add flags to specify inclusion as well as exclusion of modules/classes.
 - Allow disabling slot inheritance check.
+- Add ``--require-slots`` option.
 
 0.3.1 (2022-01-10)
 ------------------